Add route for user sign-in and implement sign-in functionality

+++ todoapi/models/dbOperation.js
@@ -1,5 +1,6 @@
 const User = require('./db/userModel');
 const connectDB = require('../config/db');
+const dbvalidation = require('./dbValidation');
 
 
 // create a class to handle all the database operations
@@ -29,7 +30,7 @@ class DbOperation {
             const user = await User.create(data);
             return this.generateResponse(user, "User created successfully");
         } catch (error) {
-            return { error: error.message };
+            return dbvalidation.generateErrorResp(error, '');
         }
     }
 
@@ -39,43 +40,75 @@ class DbOperation {
             const users = await User.find();
             return this.generateResponse(users, "Users retrieved successfully");
         } catch (error) {
-            return { error: error.message };
+            return dbvalidation.generateErrorResp(error, '');
         }
     }
 
     // Get a user by ID
     async getUserById(id) {
         try {
+            const isValidId = dbvalidation.validateId(id);
+            if (!isValidId) {
+                throw new Error("Invalid ID");
+            }
             const user = await User.findById(id);
-            if(!user) {
+            if (!user) {
                 throw new Error("User not found");
             }
             return this.generateResponse(user, "User retrieved successfully");
         } catch (error) {
-            return { error: error.message };
+            return dbvalidation.generateErrorResp(error, '');
         }
     }
 
     // Update a user by ID
     async updateUser(id, data) {
         try {
+            const isValidId = mongoose.Types.ObjectId.isValid(id);
+            if (!isValidId) {
+                throw new Error("Invalid ID");
+            }
             const user = await User.findByIdAndUpdate(id, data, {
                 new: true,
                 runValidators: true
             });
             return this.generateResponse(user, "User updated successfully");
         } catch (error) {
-            return { error: error.message }
+            return dbvalidation.generateErrorResp(error, '');
         }
     }
 
     // Delete a user by ID
     async deleteUser(id) {
         try {
+            const isValidId = mongoose.Types.ObjectId.isValid(id);
+            if (!isValidId) {
+                throw new Error("Invalid ID");
+            }
             const user = await User.findByIdAndDelete(id);
             return this.generateResponse(user, "User deleted successfully");
         } catch (error) {
-            return { error: error.message };
+            return dbvalidation.generateErrorResp(error, '');
+        }
+    }
+
+    // signin
+    async signIn(data) {
+        try {
+            const user = await User.findOne({
+                email: data.email
+            }).where('email').equals(data.email);
+            if (!user) {
+                throw new Error("User not found");
+            } else {
+                const compare = await user.isValidPassword(data.password);
+                if (!compare) {
+                    throw new Error("Invalid password");
+                }
+                return this.generateResponse(user, "User retrieved successfully");
+            }
+        } catch (error) {
+            return dbvalidation.generateErrorResp(error, '');
         }
     }
 }
